% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{benthos}
\alias{benthos}
\title{Benthos}
\format{
A \code{\link{data.frame}} with 13 columns (sites) and 92 rows (species).
}
\source{
\url{http://www.carme-n.org/?sec=data7}
}
\usage{
benthos
}
\description{
Abundances of Marine Species in Sea-Bed Samples
}
\seealso{
Other datasets: 
\code{\link{colours}},
\code{\link{countries}}
}
\concept{datasets}
\keyword{datasets}
