% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot.R
\docType{methods}
\name{biplot}
\alias{biplot}
\alias{biplot,CA-method}
\alias{biplot,PCA-method}
\title{Biplot}
\usage{
\S4method{biplot}{CA}(
  x,
  axes = c(1, 2),
  type = c("row", "column", "contributions"),
  active = TRUE,
  sup = TRUE
)

\S4method{biplot}{PCA}(
  x,
  axes = c(1, 2),
  type = c("form", "covariance"),
  active = TRUE,
  sup = TRUE
)
}
\arguments{
\item{x}{A \code{\linkS4class{CA}} or \code{\linkS4class{PCA}} object.}

\item{axes}{A length-two \code{\link{numeric}} vector giving the dimensions to be
plotted.}

\item{type}{A \code{\link{character}} string specifying the biplot to be plotted
(see below). It must be one of "\code{row}", "\code{column}", "\code{contribution}" (CA),
"\code{form}" or "\code{covariance}" (PCA).}

\item{active}{A \code{\link{logical}} scalar: should the active observations be
plotted?}

\item{sup}{A \code{\link{logical}} scalar: should the supplementary observations be
plotted?}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Biplot
}
\section{PCA Biplots}{

\describe{
\item{\code{form}}{Form biplot (row-metric-preserving).}
\item{\code{covariance}}{Covariance biplot (column-metric-preserving).}
}
}

\section{CA Biplots}{

\describe{
\item{\code{row}}{Row principal biplot.}
\item{\code{column}}{Column principal biplot.}
\item{\code{contribution}}{Contribution biplot}.
}
}

\examples{
## Replicate examples from Greenacre 2007, p. 59-68
data("countries")

## Compute principal components analysis
## All rows and all columns obtain the same weight
row_w <- rep(1 / nrow(countries), nrow(countries)) # 1/13
col_w <- rep(1 / ncol(countries), ncol(countries)) # 1/6
Y <- pca(countries, scale = FALSE, weight_row = row_w, weight_col = col_w)

## Row-metric-preserving biplot (form biplot)
biplot(Y, type = "form") +
  ggrepel::geom_label_repel()

## Column-metric-preserving biplot (covariance biplot)
biplot(Y, type = "covariance") +
  ggrepel::geom_label_repel()

## Replicate examples from Greenacre 2007, p. 79-88
data("benthos")

## Compute correspondence analysis
X <- ca(benthos)

## Row principal CA biplot
biplot(X, type = "row") +
  ggrepel::geom_label_repel()

## Column principal CA biplot
biplot(X, type = "column") +
  ggrepel::geom_label_repel()

## Contribution CA biplot
biplot(X, type = "contrib") +
  ggrepel::geom_label_repel()
}
\references{
Greenacre, M. J. \emph{Biplots in Practice}. Bilbao: Fundación BBVA, 2010.
}
\seealso{
Other plot methods: 
\code{\link{envelopes}},
\code{\link{plot_contributions}()},
\code{\link{plot_coordinates}},
\code{\link{plot_eigenvalues}}
}
\author{
N. Frerebeau
}
\concept{plot methods}
