% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{biplot_gg}
\alias{biplot_gg}
\alias{biplot_gg-method}
\title{Biplot}
\usage{
biplot_gg(x, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{CA}}, \code{\linkS4class{MCA}} or \code{\linkS4class{PCA}} object.}

\item{...}{Currently not used.}

\item{axes}{A length-two \code{\link{numeric}} vector giving the dimensions to be
plotted.}

\item{type}{A \code{\link{character}} string specifying the biplot to be plotted
(see below). It must be one of "\code{rows}", "\code{columns}", "\code{contribution}" (CA),
"\code{form}" or "\code{covariance}" (PCA). Any unambiguous substring can be given.}

\item{labels}{A \code{\link{character}} vector specifying whether
"\code{rows}"/"\code{individuals}" and/or "\code{columns}"/"\code{variables}" names must be
drawn. Any unambiguous substring can be given.}

\item{col.rows, col.columns}{A length-two \code{vector} of color specification for
the active and supplementary rows/columns.}

\item{pch.rows, pch.columns}{A length-two \code{vector} of symbol specification for
the active and supplementary rows/columns.}

\item{lty.columns}{A length-two \code{vector} of line type specification for
the active and supplementary columns.}

\item{size}{A length-two \code{\link{numeric}} vector giving range of possible sizes
(greater than 0). Only used if \code{type} is "\code{contribution}" (CA).}

\item{xlim}{A length-two \code{\link{numeric}} vector giving the x limits of the plot.
The default value, \code{NULL}, indicates that the range of the
\link[=is.finite]{finite} values to be plotted should be used.}

\item{ylim}{A length-two \code{\link{numeric}} vector giving the y limits of the plot.
The default value, \code{NULL}, indicates that the range of the
\link[=is.finite]{finite} values to be plotted should be used.}

\item{main}{A \code{\link{character}} string giving a main title for the plot.}

\item{sub}{A \code{\link{character}} string giving a subtitle for the plot.}

\item{legend}{A \code{\link{list}} of additional arguments to be passed to
\code{\link[ggplot2:theme]{ggplot2::theme()}}; names of the list are used as argument names.}
}
\value{
A \pkg{ggplot2} object.
}
\description{
Biplot
}
\seealso{
Other plot methods: 
\code{\link{biplot}()},
\code{\link{plot}()},
\code{\link{screeplot}()},
\code{\link{screeplot_gg}()},
\code{\link{viz_contributions}()},
\code{\link{viz_individuals}()},
\code{\link{viz_variables}()}
}
\author{
N. Frerebeau
}
\concept{plot methods}
