% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/tools.R
\docType{methods}
\name{burt}
\alias{burt}
\alias{burt-method}
\alias{burt,data.frame-method}
\title{Burt Table}
\usage{
burt(object)

\S4method{burt}{data.frame}(object)
}
\arguments{
\item{object}{A \code{\link{data.frame}}.}
}
\value{
A symetric \code{\link{matrix}}.
}
\description{
Computes the burt table of a factor table.
}
\examples{
## Create a factor table
x <- data.frame(
  A = c("a", "b", "a"),
  B = c("x", "y", "z")
)

## Complete disjunctive table
cdt(x)

## Burt table
burt(x)
}
\seealso{
Other tools: 
\code{\link{cdt}()}
}
\author{
N. Frerebeau
}
\concept{tools}
