% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colours}
\alias{colours}
\title{Colours}
\format{
A \code{\link{data.frame}} with 4 columns (hair colours) and 4 rows (eye
colours).
}
\source{
Lebart, L., Piron, M. and Morineau, A. \emph{Statistique exploratoire
multidimensionnelle: visualisation et inférence en fouille de données}.
Paris: Dunod, 2006, p. 170-172
}
\usage{
colours
}
\description{
Contingency table of eye and hair colours of different individuals.
}
\concept{datasets}
\keyword{datasets}
