% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ca.R
\docType{methods}
\name{correspondence}
\alias{correspondence}
\alias{ca}
\alias{ca-method}
\alias{ca,matrix-method}
\title{Correspondence Analysis}
\usage{
ca(object, ...)

\S4method{ca}{matrix}(object, n = NULL, sup_rows = NULL, sup_columns = NULL)
}
\arguments{
\item{object}{A \code{\link{matrix}} or a \code{\link{data.frame}}.}

\item{...}{Currently not used.}

\item{n}{An \code{\link{integer}} value specifying the number of dimensions
to be kept in the results (default to \code{5}).}

\item{sup_rows}{A \code{\link{numeric}} or \code{\link{logical}} vector
specifying the indices of the supplementary rows.}

\item{sup_columns}{A \code{\link{numeric}} or \code{\link{logical}} vector
specifying the indices of the supplementary columns.}
}
\value{
A \linkS4class{CA} object.
}
\description{
Computes a simple correspondence analysis based on the singular value
decomposition.
}
\examples{
## Create count data matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)
B <- matrix(data = sample(12:20, 42, TRUE), nrow = 7, ncol = 6)

## Compute correspondence analysis
X <- ca(A, sup_rows = 8:10, sup_columns = 7:10)

## Get row coordinates
get_coordinates(X, margin = 1, standard = FALSE)

## Get column coordinates
get_coordinates(X, margin = 2, standard = FALSE)

## Get row distances to centroid
get_distances(X, margin = 1)

## Get row inertias
get_inertia(X, margin = 1)

## Get row contributions
get_contributions(X, margin = 1)

## Get eigenvalues
get_eigenvalues(X)
}
\seealso{
\link{mutator}, \link{predict}

Other multivariate analysis: 
\code{\link{CA-class}},
\code{\link{predict}()}
}
\author{
N. Frerebeau
}
\concept{multivariate analysis}
