% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countries}
\alias{countries}
\title{Countries}
\format{
A \code{\link{data.frame}} with 6 columns (attributes) and 13 rows (countries).
}
\source{
Greenacre, M. J. \emph{Biplots in Practice}. Bilbao: Fundación BBVA, 2010.
}
\usage{
countries
}
\description{
Student ratings of 13 countries on six attributes.
}
\seealso{
Other datasets: 
\code{\link{benthos}},
\code{\link{colours}}
}
\concept{datasets}
\keyword{datasets}
