% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/summary.R
\docType{methods}
\name{describe}
\alias{describe}
\alias{describe,CA-method}
\alias{describe,PCA-method}
\title{Object Description}
\usage{
\S4method{describe}{CA}(x, ...)

\S4method{describe}{PCA}(x, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{CA}}, \code{\linkS4class{MCA}} or \code{\linkS4class{PCA}} object.}

\item{...}{Further parameters to be passed to \code{\link[=cat]{cat()}}.}
}
\value{
\code{describe()} is called for its side-effects. Invisibly returns \code{x}.
}
\description{
Object Description
}
\examples{
## Data from Lebart et al. 2006, p. 170-172
data("colours")

## Compute correspondence analysis
X <- ca(colours)

## Rows summary
summary(X, margin = 1)

## Columns summary
summary(X, margin = 2)
}
\seealso{
Other summary: 
\code{\link{summary}()},
\code{\link{tidy}()}
}
\author{
N. Frerebeau
}
\concept{summary}
