% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensio-defunct.R
\name{dimensio-defunct}
\alias{dimensio-defunct}
\alias{plot_rows}
\alias{plot_individuals}
\alias{plot_columns}
\alias{plot_variables}
\alias{plot_variance}
\alias{plot_contributions}
\alias{plot_cos2}
\title{Defunct Functions in dimensio}
\usage{
plot_rows(...)

plot_individuals(...)

plot_columns(...)

plot_variables(...)

plot_variance(...)

plot_contributions(...)

plot_cos2(...)
}
\arguments{
\item{...}{Not used.}
}
\description{
These functions are defunct and have been replaced.
}
\keyword{internal}
