% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensio-deprecated.R
\name{dimensio-deprecated}
\alias{dimensio-deprecated}
\title{Deprecated Functions in dimensio}
\description{
These functions still work but will be removed (defunct) in the next version.
}
\keyword{internal}
