% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensio-package.R
\docType{package}
\name{dimensio-package}
\alias{dimensio-package}
\alias{dimensio}
\title{dimensio: Multivariate Data Analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Simple Principal Components Analysis (PCA) and (Multiple) Correspondence Analysis (CA) based on the Singular Value Decomposition (SVD). This package provides S4 classes and methods to compute, extract, summarize and visualize results of multivariate data analysis. It also includes methods for partial bootstrap validation described in Greenacre (1984, ISBN: 978-0-12-299050-2) and Lebart et al. (2006, ISBN: 978-2-10-049616-7).
}
\details{
\tabular{ll}{
\strong{Version} \tab 0.14.0 \cr
\strong{License} \tab GPL-3 \cr
\strong{CRAN DOI} \tab \doi{10.32614/CRAN.package.dimensio} \cr
\strong{Zenodo DOI} \tab \doi{10.5281/zenodo.4478530} \cr
}

Archéosciences Bordeaux (UMR 6034)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\section{Package options}{

\pkg{dimensio} uses the following \code{\link[=options]{options()}} to configure behavior:
\itemize{
\item \code{dimensio.verbose}: a \code{\link{logical}} scalar. Should \R report extra
information on progress? Defaults to \code{\link[=interactive]{interactive()}}.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://codeberg.org/tesselle/dimensio}
  \item \url{https://tesselle.r-universe.dev/dimensio}
  \item \url{https://packages.tesselle.org/dimensio/}
  \item Report bugs at \url{https://codeberg.org/tesselle/dimensio/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Frerebeau \email{nicolas.frerebeau@u-bordeaux-montaigne.fr} (\href{https://orcid.org/0000-0001-5759-4944}{ORCID})

Other contributors:
\itemize{
  \item Jean-Baptiste Fourvel \email{jean-baptiste.fourvel@univ-amu.fr} (\href{https://orcid.org/0000-0002-1061-4642}{ORCID}) [contributor]
  \item Camille Thabard \email{camille.thabard@univ-tlse2.fr} (\href{https://orcid.org/0000-0002-3196-7658}{ORCID}) [contributor]
  \item Brice Lebrun (\href{https://orcid.org/0000-0001-7503-8685}{ORCID}) (Logo designer) [artist]
  \item Université Bordeaux Montaigne (\href{https://ror.org/03pbgwk21}{ROR}) [funder]
  \item CNRS (\href{https://ror.org/02feahw73}{ROR}) [funder]
}

}
\keyword{internal}
