% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutators.R
\docType{methods}
\name{dimnames}
\alias{dimnames}
\alias{dim,MultivariateAnalysis-method}
\alias{rownames,MultivariateAnalysis-method}
\alias{colnames,MultivariateAnalysis-method}
\alias{dimnames,MultivariateAnalysis-method}
\title{Dimnames of an Object}
\usage{
\S4method{dim}{MultivariateAnalysis}(x)

\S4method{rownames}{MultivariateAnalysis}(x, do.NULL = TRUE, prefix = "row")

\S4method{colnames}{MultivariateAnalysis}(x, do.NULL = TRUE, prefix = "col")

\S4method{dimnames}{MultivariateAnalysis}(x)
}
\arguments{
\item{x}{An object from which to retrieve the row or column names
(a \code{\linkS4class{CA}} or \code{\linkS4class{PCA}} object).}

\item{do.NULL}{A \code{\link{logical}} scalar. If \code{FALSE} and names are \code{NULL}, names
are created.}

\item{prefix}{A \code{\link{character}} string specifying the prefix for created names.}
}
\description{
Retrieve or set the dimnames of an object.
}
\seealso{
Other mutators: 
\code{\link{subset}()}
}
\author{
N. Frerebeau
}
\concept{mutators}
