% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/get_data.R
\docType{methods}
\name{get_data}
\alias{get_data}
\alias{get_data-method}
\alias{get_data,MultivariateAnalysis-method}
\title{Get Original Data}
\usage{
get_data(x, ...)

\S4method{get_data}{MultivariateAnalysis}(x)
}
\arguments{
\item{x}{An object from which to get element(s) (a \code{\linkS4class{CA}} or
\code{\linkS4class{PCA}} object).}

\item{...}{Currently not used.}
}
\value{
Returns a \code{\link{data.frame}} of original data.
}
\description{
Get Original Data
}
\seealso{
Other mutators: 
\code{\link{dimnames}()},
\code{\link{get_contributions}()},
\code{\link{get_coordinates}()},
\code{\link{get_distances}()},
\code{\link{get_eigenvalues}()},
\code{\link{loadings}()},
\code{\link{subset}()}
}
\author{
N. Frerebeau
}
\concept{mutators}
