% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/get_distances.R
\docType{methods}
\name{get_distances}
\alias{get_distances}
\alias{get_distances-method}
\alias{get_distances,MultivariateAnalysis-method}
\title{Get Distances}
\usage{
get_distances(x, ...)

\S4method{get_distances}{MultivariateAnalysis}(x, margin = 1)
}
\arguments{
\item{x}{An object from which to get element(s) (a \code{\linkS4class{CA}},
\code{\linkS4class{MCA}} or \code{\linkS4class{PCA}} object).}

\item{...}{Currently not used.}

\item{margin}{A length-one \code{\link{numeric}} vector giving the subscript which the
data will be returned: \code{1} indicates individuals/rows (the default), \code{2}
indicates variables/columns.}
}
\value{
A \code{\link{numeric}} vector of squared distance to the centroide.
}
\description{
Get Distances
}
\seealso{
Other mutators: 
\code{\link{dimnames}()},
\code{\link{get_contributions}()},
\code{\link{get_coordinates}()},
\code{\link{get_data}()},
\code{\link{get_eigenvalues}()},
\code{\link{loadings}()},
\code{\link{subset}()}
}
\author{
N. Frerebeau
}
\concept{mutators}
