% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutators.R
\docType{methods}
\name{groups}
\alias{groups}
\alias{get_groups}
\alias{get_groups-method}
\alias{set_groups<-}
\alias{set_groups-method}
\alias{has_groups}
\alias{has_groups-method}
\alias{get_groups,MultivariateAnalysis-method}
\alias{set_groups<-,MultivariateAnalysis-method}
\alias{set_groups,MultivariateAnalysis-method}
\alias{has_groups,MultivariateAnalysis-method}
\title{Groups}
\usage{
get_groups(x, ...)

set_groups(x) <- value

has_groups(x, ...)

\S4method{get_groups}{MultivariateAnalysis}(x, margin = 1)

\S4method{set_groups}{MultivariateAnalysis}(x) <- value

\S4method{has_groups}{MultivariateAnalysis}(x, margin = 1)
}
\arguments{
\item{x}{An object from which to get or set \code{groups}.}

\item{...}{Currently not used.}

\item{value}{A possible value for the \code{groups} of \code{x}.}

\item{margin}{A length-one \code{\link{numeric}} vector: \code{1} indicates individuals/rows
(the default), \code{2} indicates variables/columns.}
}
\value{
\itemize{
\item \code{set_groups()} returns an object of the same sort as \code{x} with the new
group names assigned.
\item \code{get_groups()} returns a \code{\link{character}} vector giving the group names of \code{x}.
\item \code{has_groups()} returns a \code{\link{logical}} scalar specifying whether or not \code{x}
has groups.
}
}
\description{
Retrieves or defines the groups to which the observations belong.
}
\seealso{
Other mutators: 
\code{\link{dimnames}()},
\code{\link{loadings}()},
\code{\link{subset}()}
}
\author{
N. Frerebeau
}
\concept{mutators}
