% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/loadings.R
\docType{methods}
\name{loadings}
\alias{loadings}
\alias{loadings,PCA-method}
\title{Extract Loadings}
\usage{
\S4method{loadings}{PCA}(x, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{PCA}} object.}

\item{...}{Currently not used.}
}
\value{
Returns variable loadings (i.e. the coefficients of the linear combination
of the original variables).
}
\description{
Extract loadingsin principal components analysis.
}
\note{
\code{loadings()} is only implemented for consistency with \code{\link[stats:loadings]{stats::loadings()}}.
}
\seealso{
Other mutators: 
\code{\link{dimnames}()},
\code{\link{get_contributions}()},
\code{\link{get_coordinates}()},
\code{\link{get_data}()},
\code{\link{get_distances}()},
\code{\link{get_eigenvalues}()},
\code{\link{subset}()}
}
\author{
N. Frerebeau
}
\concept{mutators}
