% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MultivariateAnalysis-class}
\alias{MultivariateAnalysis-class}
\alias{.MultivariateAnalysis}
\title{Output of Multivariate Data Analysis}
\description{
A virtual S4 class to store the output of a multivariate data analysis.
}
\section{Slots}{

\describe{
\item{\code{data}}{A \code{\link{numeric}} \code{\link{matrix}}.}

\item{\code{dimension}}{An \code{\link{integer}} giving the dimension of the
solution.}

\item{\code{singular_values}}{A \code{\link{numeric}} vector giving the singular
values.}

\item{\code{rows}}{A \linkS4class{MultivariateResults} object.}

\item{\code{columns}}{A \linkS4class{MultivariateResults} object.}
}}

\section{Subset}{

In the code snippets below, \code{x} is a \code{MultivariateAnalysis}
object.
\describe{
\item{\code{x[[i]]}}{Extracts information from a slot selected by
subscript \code{i}. \code{i} is a length-one \code{\link{character}}
vector.}
}
}

\seealso{
Other class: 
\code{\link{CA-class}},
\code{\link{MultivariateResults-class}},
\code{\link{MultivariateSummary}},
\code{\link{PCA-class}}
}
\author{
N. Frerebeau
}
\concept{class}
