% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MultivariateBootstrap}
\alias{MultivariateBootstrap}
\alias{.MultivariateBootstrap}
\alias{MultivariateBootstrap-class}
\title{Output of Bootstrap Replications}
\description{
A virtual S4 class to store the output of a bootstrap analysis.
}
\section{Slots}{

\describe{
\item{\code{replications}}{An \code{\link{integer}} giving the number of bootstrap
replications.}
}}

\seealso{
Other class: 
\code{\link{CA-class}},
\code{\link{MultivariateAnalysis}},
\code{\link{MultivariateResults}},
\code{\link{MultivariateSummary}},
\code{\link{PCA-class}}
}
\author{
N. Frerebeau
}
\concept{class}
