% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MultivariateResults-class}
\alias{MultivariateResults-class}
\alias{.MultivariateResults}
\title{Multivariate Data Analysis Results}
\description{
An S4 class to store the results of a multivariate data analysis.
}
\section{Slots}{

\describe{
\item{\code{names}}{A \code{\link{character}} vector specifying the row names.}

\item{\code{coordinates}}{A \code{\link{numeric}} \code{\link{matrix}}
giving the principal coordinates.}

\item{\code{standard}}{A \code{\link{numeric}} \code{\link{matrix}}
giving the standard coordinates.}

\item{\code{contributions}}{A \code{\link{numeric}} \code{\link{matrix}}
giving the contributions to the definition of the dimensions.}

\item{\code{cosine}}{A \code{\link{numeric}} \code{\link{matrix}}
giving the \eqn{cos^2}{cos2} values.}

\item{\code{distances}}{A \code{\link{numeric}} vector giving the distances to
centroid.}

\item{\code{weights}}{A \code{\link{numeric}} vector giving the masses/weights.}

\item{\code{supplement}}{A \code{\link{logical}} vector specifying the supplementary
points.}
}}

\seealso{
Other multivariate analysis: 
\code{\link{CA-class}},
\code{\link{MultivariateAnalysis-class}},
\code{\link{PCA-class}},
\code{\link{ca}()},
\code{\link{pca}()},
\code{\link{predict}()}
}
\author{
N. Frerebeau
}
\concept{multivariate analysis}
