% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MultivariateSummary}
\alias{MultivariateSummary}
\alias{.MultivariateSummary}
\alias{MultivariateSummary-class}
\alias{SummaryCA-class}
\alias{.SummaryCA}
\alias{SummaryPCA-class}
\alias{.SummaryPCA}
\title{Summary of Multivariate Data Analysis}
\description{
A virtual S4 class to store the summary of a multivariate data analysis.
}
\section{Slots}{

\describe{
\item{\code{data}}{A \code{\link{numeric}} \code{\link{matrix}}.}

\item{\code{eigenvalues}}{A \code{\link{numeric}} \code{\link{matrix}}.}

\item{\code{results}}{A \code{\link{numeric}} \code{\link{matrix}}.}

\item{\code{supplement}}{A \code{\link{logical}} vector specifying the supplementary points.}

\item{\code{margin}}{An \code{\link{integer}}.}
}}

\seealso{
Other class: 
\code{\link{CA-class}},
\code{\link{MultivariateAnalysis}},
\code{\link{MultivariateBootstrap}},
\code{\link{MultivariateResults}},
\code{\link{PCA-class}}
}
\author{
N. Frerebeau
}
\concept{class}
\keyword{internal}
