% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutators.R, R/pca.R
\docType{methods}
\name{mutators}
\alias{mutators}
\alias{dim,MultivariateAnalysis-method}
\alias{rownames,MultivariateAnalysis-method}
\alias{colnames,MultivariateAnalysis-method}
\alias{dimnames,MultivariateAnalysis-method}
\alias{loadings,PCA-method}
\title{Get Results}
\usage{
\S4method{dim}{MultivariateAnalysis}(x)

\S4method{rownames}{MultivariateAnalysis}(x)

\S4method{colnames}{MultivariateAnalysis}(x)

\S4method{dimnames}{MultivariateAnalysis}(x)

\S4method{loadings}{PCA}(x)
}
\arguments{
\item{x}{An object from which to get element(s) (a \code{\linkS4class{CA}} or
\code{\linkS4class{PCA}} object).}
}
\value{
\code{loadings()} returns variable loadings (i.e. the coefficients of the linear
combination of the original variables). \code{loadings()} is only implemented for
consistency with \pkg{[stats][stats::loadings]}.
}
\description{
Getters to retrieve parts of an object.
}
\seealso{
Other mutators: 
\code{\link{get_contributions}()},
\code{\link{get_coordinates}()},
\code{\link{get_data}()},
\code{\link{get_distances}()},
\code{\link{get_eigenvalues}()},
\code{\link{subset}()}
}
\author{
N. Frerebeau
}
\concept{mutators}
