% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/pca.R
\docType{methods}
\name{pca}
\alias{pca}
\alias{pca-method}
\alias{pca,data.frame-method}
\alias{pca,matrix-method}
\title{Principal Components Analysis}
\usage{
pca(object, ...)

\S4method{pca}{data.frame}(
  object,
  center = TRUE,
  scale = TRUE,
  rank = NULL,
  sup_ind = NULL,
  sup_var = NULL,
  weight_ind = NULL,
  weight_var = NULL
)

\S4method{pca}{matrix}(
  object,
  center = TRUE,
  scale = TRUE,
  rank = NULL,
  sup_ind = NULL,
  sup_var = NULL,
  weight_ind = NULL,
  weight_var = NULL
)
}
\arguments{
\item{object}{A \eqn{m \times p}{m x p} numeric \code{\link{matrix}} or a
\code{\link{data.frame}}.}

\item{...}{Currently not used.}

\item{center}{A \code{\link{logical}} scalar: should the variables be shifted
to be zero centered?}

\item{scale}{A \code{\link{logical}} scalar: should the variables be scaled
to unit variance?}

\item{rank}{An \code{\link{integer}} value specifying the maximal number of
components to be kept in the results. If \code{NULL} (the default),
\eqn{p - 1} components will be returned.}

\item{sup_ind}{A \code{\link{numeric}} or \code{\link{logical}} vector
specifying the indices of the supplementary individuals.}

\item{sup_var}{A \code{\link{numeric}} or \code{\link{logical}} vector
specifying the indices of the supplementary variables.}

\item{weight_ind}{A \code{\link{numeric}} vector specifying the active
individual weights. If \code{NULL} (the default), no weights are used.}

\item{weight_var}{A \code{\link{numeric}} vector specifying the active
individual weights. If \code{NULL} (the default), no weights are used.}
}
\value{
A \linkS4class{PCA} object.
}
\description{
Computes a principal components analysis based on the singular value
decomposition.
}
\examples{
## Load data
library(codex)
data("compiegne")

## Compute principal components analysis
X <- pca(compiegne, scale = TRUE, sup_var = 7:10)

## Get row coordinates
get_coordinates(X, margin = 1)

## Get column coordinates
get_coordinates(X, margin = 2)

## Get row contributions
get_contributions(X, margin = 1)

## Get correlations between variables and dimensions
get_correlations(X)

## Get eigenvalues
get_eigenvalues(X)
}
\references{
Lebart, L., Piron, M. and Morineau, A. \emph{Statistique exploratoire
multidimensionnelle: visualisation et inférence en fouille de données}.
Paris: Dunod, 2006.
}
\seealso{
\link[=mutator]{get_*()}, \link[=predict]{predict()},
\link[=svd]{svd()}

Other multivariate analysis: 
\code{\link{bootstrap}()},
\code{\link{ca}()},
\code{\link{predict}()}
}
\author{
N. Frerebeau
}
\concept{multivariate analysis}
