% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot.R
\docType{methods}
\name{plot_coordinates}
\alias{plot_coordinates}
\alias{plot_coordinates-method}
\alias{plot_individuals}
\alias{plot_individuals-method}
\alias{plot_variables}
\alias{plot_variables-method}
\alias{plot_coordinates,CA-method}
\alias{plot_coordinates,PCA-method}
\alias{plot_individuals,PCA-method}
\alias{plot_variables,PCA-method}
\title{Visualize Factor Map}
\usage{
plot_coordinates(object, ...)

plot_individuals(object, ...)

plot_variables(object, ...)

\S4method{plot_coordinates}{CA}(
  object,
  margin = c(1, 2),
  axes = c(1, 2),
  active = TRUE,
  sup = TRUE,
  highlight = NULL,
  group = NULL
)

\S4method{plot_coordinates}{PCA}(
  object,
  margin = 1,
  axes = c(1, 2),
  active = TRUE,
  sup = TRUE,
  highlight = NULL,
  group = NULL
)

\S4method{plot_individuals}{PCA}(
  object,
  axes = c(1, 2),
  active = TRUE,
  sup = TRUE,
  highlight = NULL,
  group = NULL
)

\S4method{plot_variables}{PCA}(
  object,
  axes = c(1, 2),
  active = TRUE,
  sup = TRUE,
  highlight = NULL,
  group = NULL
)
}
\arguments{
\item{object}{A \linkS4class{CA} or \linkS4class{PCA} object.}

\item{...}{Currently not used.}

\item{margin}{A length-one \code{\link{numeric}} vector giving the subscript
which the data will be returned: \code{1} indicates individuals/rows (the
default), \code{2} indicates variables/columns, \code{c(1, 2)} indicates
rows and columns (\code{CA}).}

\item{axes}{A length-two \code{\link{numeric}} vector giving the dimensions
to be plotted.}

\item{active}{A \code{\link{logical}} scalar: should the active
observations be plotted?}

\item{sup}{A \code{\link{logical}} scalar: should the supplementary
observations be plotted?}

\item{highlight}{A \code{\link{character}} string giving XXX.
It must be one of "\code{coordinates}", "\code{contributions}",
"\code{cos2}" or "\code{distances}". Any unambiguous substring can be given.
If \code{NULL} (the default), no highlighting is applied.}

\item{group}{A vector of categories specifying the categorical
variable from which to color the individuals (see details).}
}
\description{
Plots factor map.
}
\details{
Point shapes and line types are set whether an observation is a
row/individual or a column/variable and is active or supplementary.

Colours are set according to \code{highlight} or \code{group}
(if both are \code{NULL}, the same rule as for shapes is used).
}
\examples{
## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 20)

## Compute correspondence analysis
X <- ca(A, sup_row = 11:15)

## Plot observations
plot_coordinates(X)
}
\seealso{
Other visualization: 
\code{\link{plot_contributions}()},
\code{\link{plot_eigenvalues}()}
}
\author{
N. Frerebeau
}
\concept{visualization}
