% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensio-internal.R
\name{prepare_legend}
\alias{prepare_legend}
\title{Build a Legend}
\usage{
prepare_legend(x, args, points = TRUE, lines = TRUE)
}
\arguments{
\item{x}{A \code{\link{data.frame}} returned by \code{\link[=prepare]{prepare()}}.}

\item{args}{A \code{\link{list}} of additional arguments to be passed to
\code{\link[graphics:legend]{graphics::legend()}}; names of the list are used as argument names.
If \code{NULL}, no legend is displayed.}

\item{points}{A \code{\link{logical}} scalar: legend for points?}

\item{lines}{A \code{\link{logical}} scalar: legend for lines?}
}
\description{
Build a Legend
}
\author{
N. Frerebeau
}
\keyword{internal}
