% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{screeplot_gg}
\alias{screeplot_gg}
\alias{screeplot_gg-method}
\title{Scree Plot}
\usage{
screeplot_gg(x, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{CA}}, \code{\linkS4class{MCA}} or \code{\linkS4class{PCA}} object.}

\item{...}{Extra parameters to be passed to \code{\link[ggplot2:geom_bar]{ggplot2::geom_col()}} or
\code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}.}

\item{eigenvalues}{A \code{\link{logical}} scalar: should the eigenvalues be plotted
instead of variance/inertia?}

\item{cumulative}{A \code{\link{logical}} scalar: should the cumulative percentages of
variance be plotted?}

\item{labels}{A \code{\link{logical}} scalar: should text labels be drawn on top of
bars?}

\item{limit}{An \code{\link{integer}} specifying the number of top elements to be
displayed.}
}
\value{
A \pkg{ggplot2} object.
}
\description{
Plot eigenvalues (scree plot) or variances histogram.
}
\examples{
## Load data
data("iris")

## Compute principal components analysis
X <- pca(iris, scale = TRUE, sup_quali = "Species")

## Screeplot
screeplot(X)
screeplot(X, cumulative = TRUE)
}
\seealso{
Other plot methods: 
\code{\link{biplot}()},
\code{\link{biplot_gg}()},
\code{\link{plot}()},
\code{\link{screeplot}()},
\code{\link{viz_contributions}()},
\code{\link{viz_individuals}()},
\code{\link{viz_variables}()}
}
\author{
N. Frerebeau
}
\concept{plot methods}
