% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot.R
\docType{methods}
\name{viz_contributions}
\alias{viz_contributions}
\alias{viz_contributions-method}
\alias{viz_cos2}
\alias{viz_cos2-method}
\alias{viz_contributions,MultivariateAnalysis-method}
\alias{viz_cos2,MultivariateAnalysis-method}
\title{Visualize Contributions and cos2}
\usage{
viz_contributions(x, ...)

viz_cos2(x, ...)

\S4method{viz_contributions}{MultivariateAnalysis}(
  x,
  margin = 2,
  axes = 1,
  sort = TRUE,
  decreasing = TRUE,
  limit = 10,
  horiz = FALSE,
  col = "grey90",
  border = "grey10",
  ...
)

\S4method{viz_cos2}{MultivariateAnalysis}(
  x,
  margin = 2,
  axes = c(1, 2),
  active = TRUE,
  sup = TRUE,
  sort = TRUE,
  decreasing = TRUE,
  limit = 10,
  horiz = FALSE,
  col = "grey90",
  border = "grey10",
  ...
)
}
\arguments{
\item{x}{A \code{\linkS4class{CA}} or \code{\linkS4class{PCA}} object.}

\item{...}{Extra parameters to be passed to \code{\link[graphics:barplot]{graphics::barplot()}}.}

\item{margin}{A length-one \code{\link{numeric}} vector giving the subscript which the
data will be returned: \code{1} indicates individuals/rows (the default), \code{2}
indicates variables/columns.}

\item{axes}{A length-one \code{\link{numeric}} vector giving the dimensions to be
plotted.}

\item{sort}{A \code{\link{logical}} scalar: should the data be sorted?}

\item{decreasing}{A \code{\link{logical}} scalar: should the sort order be decreasing?
Only used if \code{sort} is \code{TRUE}.}

\item{limit}{An \code{\link{integer}} specifying the number of top elements to be
displayed.}

\item{horiz}{A \code{\link{logical}} scalar: should the bars be drawn horizontally
with the first at the bottom?}

\item{col, border}{A \code{\link{character}} string specifying the bars infilling and
border colors.}

\item{active}{A \code{\link{logical}} scalar: should the active observations be
plotted?}

\item{sup}{A \code{\link{logical}} scalar: should the supplementary observations be
plotted?}
}
\value{
\code{viz_contributions()} and \code{viz_cos2()} are called for their side-effects:
they result in a graphic being displayed. Invisibly return \code{x}.
}
\description{
Plots contributions histogram and \eqn{cos^2}{cos2} scatterplot.
}
\examples{
## Load data
data("iris")

## Compute principal components analysis
X <- pca(iris, scale = TRUE)

## Get row contributions
head(get_contributions(X, margin = 1))

## Plot contributions
viz_contributions(X)

## Plot cos2
viz_cos2(X)
}
\seealso{
Other plot methods: 
\code{\link{biplot}()},
\code{\link{screeplot}()},
\code{\link{viz_coordinates}},
\code{\link{viz_wrap}},
\code{\link{wrap}}
}
\author{
N. Frerebeau
}
\concept{plot methods}
