% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{viz_contributions_gg}
\alias{viz_contributions_gg}
\alias{viz_contributions_gg-method}
\title{Visualize Contributions and cos2}
\usage{
viz_contributions_gg(x, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{CA}}, \code{\linkS4class{MCA}} or \code{\linkS4class{PCA}} object.}

\item{...}{Extra parameters to be passed to \code{\link[ggplot2:geom_bar]{ggplot2::geom_col()}}.}

\item{margin}{A length-one \code{\link{numeric}} vector giving the subscript which the
data will be returned: \code{1} indicates individuals/rows (the default), \code{2}
indicates variables/columns.}

\item{axes}{A \code{\link{numeric}} vector giving the dimensions to be plotted.}

\item{sort}{A \code{\link{logical}} scalar: should the data be sorted?}

\item{decreasing}{A \code{\link{logical}} scalar: should the sort order be decreasing?
Only used if \code{sort} is \code{TRUE}.}

\item{limit}{An \code{\link{integer}} specifying the number of top elements to be
displayed.}

\item{horiz}{A \code{\link{logical}} scalar: should the bars be drawn horizontally
with the first at the bottom?}

\item{col, border}{A \code{\link{character}} string specifying the bars infilling and
border colors.}
}
\value{
A \pkg{ggplot2} object.
}
\description{
Plots contributions histogram.
}
\details{
The red dashed line indicates the expected average contribution (variables
with a contribution larger than this cutoff can be considered as important
in contributing to the component).
}
\examples{
## Load data
data("iris")

## Compute principal components analysis
X <- pca(iris, scale = TRUE, sup_quali = "Species")

## Plot contributions using ggplot2
viz_contributions_gg(X, axes = 1)
}
\seealso{
Other plot methods: 
\code{\link{biplot}()},
\code{\link{biplot_gg}()},
\code{\link{plot}()},
\code{\link{screeplot}()},
\code{\link{screeplot_gg}()},
\code{\link{viz_contributions}()},
\code{\link{viz_individuals}()},
\code{\link{viz_variables}()},
\code{\link{viz_variables_gg}()}
}
\author{
N. Frerebeau
}
\concept{plot methods}
