% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/viz_coordinates.R
\docType{methods}
\name{viz_individuals}
\alias{viz_individuals}
\alias{viz_individuals-method}
\alias{viz_rows}
\alias{viz_rows-method}
\alias{viz_rows,MultivariateAnalysis-method}
\alias{viz_rows,BootstrapCA-method}
\alias{viz_individuals,PCA-method}
\title{Visualize Individuals Factor Map}
\usage{
viz_individuals(x, ...)

viz_rows(x, ...)

\S4method{viz_rows}{MultivariateAnalysis}(
  x,
  axes = c(1, 2),
  active = TRUE,
  sup = TRUE,
  labels = FALSE,
  alpha = NULL,
  colour = NULL,
  shape = NULL,
  size = NULL,
  main = NULL,
  sub = NULL,
  ...
)

\S4method{viz_rows}{BootstrapCA}(x, axes = c(1, 2), ...)

\S4method{viz_individuals}{PCA}(
  x,
  axes = c(1, 2),
  active = TRUE,
  sup = TRUE,
  labels = FALSE,
  alpha = NULL,
  colour = NULL,
  shape = NULL,
  size = NULL,
  main = NULL,
  sub = NULL,
  ...
)
}
\arguments{
\item{x}{A \code{\linkS4class{CA}} or \code{\linkS4class{PCA}} object.}

\item{...}{Further \link[graphics:par]{graphical parameters} (see below).}

\item{axes}{A length-two \code{\link{numeric}} vector giving the dimensions to be
plotted.}

\item{active}{A \code{\link{logical}} scalar: should the active observations be
plotted?}

\item{sup}{A \code{\link{logical}} scalar: should the supplementary observations be
plotted?}

\item{labels}{A \code{\link{logical}} scalar: should labels be drawn?}

\item{alpha, colour, shape, size}{A vector specifying the
information to be highlighted (will be mapped to the corresponding
aesthetic; see examples and vignettes).
If a single \code{character} string is passed, it must be one of "\code{observation}",
"\code{mass}", "\code{sum}", "\code{contribution}" or "\code{cos2}" (see details).
Any unambiguous substring can be given.
If \code{NULL} (the default), no highlighting is applied.}

\item{main}{A \code{\link{character}} string giving a main title for the plot.}

\item{sub}{A \code{\link{character}} string giving a subtitle for the plot.}
}
\value{
\verb{viz_*()} is called for its side-effects: it results in a graphic
being displayed. Invisibly returns \code{x}.
}
\description{
Plots row/individual principal coordinates.
}
\details{
Available statistics:
\describe{
\item{\code{observation}}{Whether an observation is active or supplementary.}
\item{\code{mass}}{Weight/mass of each observation.}
\item{\code{sum}}{Sum of squared coordinates along \code{axes}.}
\item{\code{contribution}}{Joint contributions to the definition of \code{axes}.}
\item{\code{cos2}}{Joint \eqn{cos^2}{cos2} along \code{axes}.}
}

Commonly used \link[graphics:par]{graphical parameters} are:
\describe{
\item{\code{col}}{The colors for lines and points. Multiple colors can be
specified so that each point can be given its own color.}
\item{\code{pch}}{A vector of plotting characters or symbols.}
\item{\code{cex}}{A \code{numeric} vector giving the amount by which plotting
characters and symbols should be scaled relative to the
default.}
\item{\code{lty}}{A vector of line types.}
\item{\code{lwd}}{A vector of line widths.}
}
}
\examples{
## Load data
data("iris")

## Compute principal components analysis
X <- pca(iris, scale = TRUE)

## Plot individuals
viz_individuals(X, colour = iris$Species, shape = iris$Species)

viz_individuals(X, colour = iris$Sepal.Width, size = "cos2")

viz_individuals(X, colour = "contribution", size = "contribution", pch = 16)

## Plot variables
viz_variables(X)

viz_variables(X, colour = "contribution")
}
\seealso{
Other plot methods: 
\code{\link{biplot}()},
\code{\link{screeplot}()},
\code{\link{viz_contributions}()},
\code{\link{viz_variables}()},
\code{\link{viz_wrap}},
\code{\link{wrap}}
}
\author{
N. Frerebeau
}
\concept{plot methods}
