% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_coordinates.R
\name{viz_legend}
\alias{viz_legend}
\title{Add Legend}
\usage{
viz_legend(x, args, points = TRUE, lines = TRUE)
}
\arguments{
\item{x}{A \code{\link{data.frame}} returned by \code{\link[=prepare_plot]{prepare_plot()}}.}

\item{args}{A \code{\link{list}} of additional arguments to be passed to
\code{\link[graphics:legend]{graphics::legend()}}; names of the list are used as argument names.
If \code{NULL} or empty, no legend is displayed.}

\item{points}{A \code{\link{logical}} scalar: legend for points?}

\item{lines}{A \code{\link{logical}} scalar: legend for lines?}
}
\description{
Add Legend
}
\author{
N. Frerebeau
}
\keyword{internal}
