% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{viz_variables_gg}
\alias{viz_variables_gg}
\alias{viz_variables_gg-method}
\title{Visualize Variables Factor Map}
\usage{
viz_variables_gg(x, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{CA}}, \code{\linkS4class{MCA}} or \code{\linkS4class{PCA}} object.}

\item{...}{Further parameters to be passed to \code{\link[ggplot2:geom_segment]{ggplot2::geom_segment()}}
(if \code{x} is a \code{\linkS4class{PCA}} object) or \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}
(if \code{x} is a \code{\linkS4class{CA}} object).}
}
\value{
A \pkg{ggplot2} object.
}
\description{
Plots column/variable principal coordinates.
}
\examples{
## Load data
data("iris")

## Compute principal components analysis
X <- pca(iris, scale = TRUE, sup_quali = "Species")

## Plot variables with ggplot2
viz_variables_gg(X)

## Graphical parameters
## Continuous values
viz_variables_gg(X, extra_quanti = "contribution", size = c(0, 1))

## Discrete values
viz_variables_gg(X, extra_quali = c("Petal", "Petal", "Sepal", "Sepal"),
                 symbol = c(1, 3))
}
\seealso{
Other plot methods: 
\code{\link{biplot}()},
\code{\link{biplot_gg}()},
\code{\link{plot}()},
\code{\link{screeplot}()},
\code{\link{screeplot_gg}()},
\code{\link{viz_contributions}()},
\code{\link{viz_individuals}()},
\code{\link{viz_variables}()}
}
\author{
N. Frerebeau
}
\concept{plot methods}
