% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CA-class}
\alias{CA-class}
\alias{.CA}
\title{CA Results}
\description{
An S4 class to store the results of a simple correspondence analysis.
}
\examples{
## Load data
library(codex)
data("compiegne")

## The chi square of independence between the two variables
stats::chisq.test(compiegne)

## Compute correspondence analysis
X <- ca(compiegne, sup_col = 7:10)

## Get row coordinates
get_coordinates(X, margin = 1)

## Get column coordinates
get_coordinates(X, margin = 2)

## Get row distances to centroid
get_distances(X, margin = 1)

## Get row inertias
get_inertia(X, margin = 1)

## Get row contributions
get_contributions(X, margin = 1)

## Get eigenvalues
get_eigenvalues(X)
}
\seealso{
Other multivariate analysis: 
\code{\link{MultivariateAnalysis-class}},
\code{\link{MultivariateResults-class}},
\code{\link{MultivariateSummary}},
\code{\link{PCA-class}},
\code{\link{ca}()},
\code{\link{pca}()},
\code{\link{predict}()}
}
\author{
N. Frerebeau
}
\concept{multivariate analysis}
