% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CA-class}
\alias{CA-class}
\alias{.CA}
\title{Correspondence Analysis Results}
\description{
An S4 class to store the results of a simple correspondence analysis.
}
\section{Slots}{

\describe{
\item{\code{data}}{A \code{\link{numeric}} \code{\link{matrix}}}

\item{\code{dimension}}{An \code{\link{integer}} giving the dimension of the
solution.}

\item{\code{row_names}}{A \code{\link{character}} vector specifying the row names.}

\item{\code{row_coordinates}}{A \code{\link{numeric}} \code{\link{matrix}}
giving the row standard coordinates.}

\item{\code{row_distances}}{A \code{\link{numeric}} vector giving the row chi-square
distances to centroid.}

\item{\code{row_inertias}}{A \code{\link{numeric}} vector giving the row inertias.}

\item{\code{row_masses}}{A \code{\link{numeric}} vector giving the row masses.}

\item{\code{row_supplement}}{A \code{\link{logical}} vector specifying if a row is a
supplementary point.}

\item{\code{column_names}}{A \code{\link{character}} vector specifying the column
names.}

\item{\code{column_coordinates}}{A \code{\link{numeric}} \code{\link{matrix}}
giving the column standard coordinates.}

\item{\code{column_distances}}{A \code{\link{numeric}} vector giving the column
chi-square distances to centroid.}

\item{\code{column_inertias}}{A \code{\link{numeric}} vector giving the column
inertias.}

\item{\code{column_masses}}{A \code{\link{numeric}} vector giving the column masses.}

\item{\code{column_supplement}}{A \code{\link{logical}} vector specifying if a column
is a supplementary point.}

\item{\code{singular_values}}{A \code{\link{numeric}} vector giving the singular
values.}
}}

\section{Subset}{

In the code snippets below, \code{x} is a \code{CA} object.
\describe{
\item{\code{x[[i]]}}{Extracts informations from a slot selected by
subscript \code{i}. \code{i} is a length-one \code{\link{character}}
vector. Returns the corresponding slot values.}
}
}

\examples{
## Create count data matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)
B <- matrix(data = sample(12:20, 42, TRUE), nrow = 7, ncol = 6)

## Compute correspondence analysis
X <- ca(A, sup_rows = 8:10, sup_columns = 7:10)

## Get row coordinates
get_coordinates(X, margin = 1, standard = FALSE)

## Get column coordinates
get_coordinates(X, margin = 2, standard = FALSE)

## Get row distances to centroid
get_distances(X, margin = 1)

## Get row inertias
get_inertia(X, margin = 1)

## Get row contributions
get_contributions(X, margin = 1)

## Get eigenvalues
get_eigenvalues(X)
}
\seealso{
Other multivariate analysis: 
\code{\link{correspondence}},
\code{\link{predict}()}
}
\author{
N. Frerebeau
}
\concept{multivariate analysis}
