% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensio-deprecated.R
\name{dimensio-deprecated}
\alias{dimensio-deprecated}
\alias{stat_hull}
\alias{plot_rows}
\alias{plot_rows,MultivariateAnalysis-method}
\alias{plot_columns}
\alias{plot_columns,MultivariateAnalysis-method}
\alias{plot_columns,BootstrapPCA-method}
\alias{plot_individuals}
\alias{plot_individuals,PCA-method}
\alias{plot_variables}
\alias{plot_variables,PCA-method}
\alias{plot_variance}
\alias{plot_variance,MultivariateAnalysis-method}
\alias{plot_contributions}
\alias{plot_contributions,MultivariateAnalysis-method}
\alias{plot_cos2}
\alias{plot_cos2,MultivariateAnalysis-method}
\title{Deprecated Functions in dimensio}
\usage{
stat_hull(
  mapping = NULL,
  data = NULL,
  geom = "polygon",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)

plot_rows(object, ...)

\S4method{plot_rows}{MultivariateAnalysis}(
  object,
  axes = c(1, 2),
  active = TRUE,
  sup = TRUE,
  alpha = NULL,
  colour = NULL,
  fill = NULL,
  shape = NULL,
  size = NULL,
  group = NULL
)

plot_columns(object, ...)

\S4method{plot_columns}{MultivariateAnalysis}(
  object,
  axes = c(1, 2),
  active = TRUE,
  sup = TRUE,
  alpha = NULL,
  colour = NULL,
  fill = NULL,
  shape = NULL,
  size = NULL,
  group = NULL
)

\S4method{plot_columns}{BootstrapPCA}(
  object,
  axes = c(1, 2),
  active = TRUE,
  sup = TRUE,
  alpha = NULL,
  colour = NULL,
  fill = NULL,
  shape = NULL,
  size = NULL,
  group = NULL
)

plot_individuals(object, ...)

\S4method{plot_individuals}{PCA}(
  object,
  axes = c(1, 2),
  active = TRUE,
  sup = TRUE,
  alpha = NULL,
  colour = NULL,
  fill = NULL,
  shape = NULL,
  size = NULL,
  group = NULL
)

plot_variables(object, ...)

\S4method{plot_variables}{PCA}(
  object,
  axes = c(1, 2),
  active = TRUE,
  sup = TRUE,
  alpha = NULL,
  colour = NULL,
  linetype = NULL,
  size = NULL,
  group = NULL
)

plot_variance(object, ...)

\S4method{plot_variance}{MultivariateAnalysis}(
  object,
  select = NULL,
  variance = TRUE,
  cumulative = TRUE,
  fill = "grey30",
  border = "grey10",
  colour = "red"
)

plot_contributions(object, ...)

\S4method{plot_contributions}{MultivariateAnalysis}(
  object,
  margin = 2,
  axes = 1,
  sort = TRUE,
  decreasing = TRUE,
  limit = 10,
  fill = "grey30",
  border = "grey10"
)

plot_cos2(object, ...)

\S4method{plot_cos2}{MultivariateAnalysis}(
  object,
  margin = 2,
  axes = c(1, 2),
  active = TRUE,
  sup = TRUE,
  sort = TRUE,
  decreasing = TRUE,
  limit = 10,
  fill = "grey30",
  border = "grey10"
)
}
\description{
These functions still work but will be removed (defunct) in the next version.
}
\keyword{internal}
