% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensio-package.R
\docType{package}
\name{dimensio-package}
\alias{dimensio-package}
\alias{_PACKAGE}
\alias{dimensio}
\title{dimensio: Multivariate Data Analysis}
\description{
Simple Principal Components Analysis (PCA) and
    Correspondence Analysis (CA). This package provides classes and methods to
    compute, extract, summarize and visualize results of multivariate data
    analysis.
}
\details{
\tabular{ll}{
\strong{Package:} \tab dimensio \cr
\strong{Type:} \tab Package \cr
\strong{Version:} \tab 1.0.0 \cr
\strong{License:} \tab GPL-3 \cr
\strong{DOI:} \tab \href{https://doi.org/xxx}{xxx}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{http://dimensio.archaeo.science}
  \item \url{https://github.com/nfrerebeau/dimensio}
  \item Report bugs at \url{https://github.com/nfrerebeau/dimensio/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\tabular{ll}{
Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
Brice Lebrun \tab \emph{Université Bordeaux Montaigne, France} \cr
}

\strong{Package maintainer}

Nicolas Frerebeau\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}

IRAMAT-CRP2A (UMR 5060)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\keyword{internal}
