% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensio-package.R
\docType{package}
\name{dimensio-package}
\alias{dimensio-package}
\alias{_PACKAGE}
\alias{dimensio}
\title{dimensio: Multivariate Data Analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Simple Principal Components Analysis (PCA) and Correspondence Analysis (CA) based on the Singular Value Decomposition (SVD). This package provides S4 classes and methods to compute, extract, summarize and visualize results of multivariate data analysis. It also includes methods for partial bootstrap validation described in Greenacre (1984) <isbn:978-0-12-299050-2> and Lebart et al. (2006) <isbn:978-2-10-049616-7>.
}
\details{
\tabular{ll}{
\strong{Package:} \tab dimensio \cr
\strong{Type:} \tab Package \cr
\strong{Version:} \tab 0.4.0 \cr
\strong{License:} \tab GPL-3 \cr
\strong{Zenodo:} \tab \doi{10.5281/zenodo.4478530} \cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://packages.tesselle.org/dimensio/}
  \item \url{https://github.com/tesselle/dimensio}
  \item Report bugs at \url{https://github.com/tesselle/dimensio/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\tabular{ll}{
Jean-Baptiste Fourvel \tab \emph{CNRS, France} \cr
Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
Brice Lebrun \tab \emph{Université Bordeaux Montaigne, France} \cr
}

\strong{Package maintainer}

Nicolas Frerebeau\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}

Archéosciences Bordeaux (UMR 6034)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\keyword{internal}
