% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MCA-class}
\alias{MCA-class}
\alias{.MCA}
\title{MCA Results}
\description{
An S4 class to store the results of a multiple correspondence analysis.
}
\seealso{
Other class: 
\code{\link{CA-class}},
\code{\link{MultivariateAnalysis}},
\code{\link{MultivariateBootstrap}},
\code{\link{MultivariateResults}},
\code{\link{MultivariateSummary}},
\code{\link{PCA-class}}
}
\author{
N. Frerebeau
}
\concept{class}
\keyword{internal}
