% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ca.R
\docType{methods}
\name{mutator}
\alias{mutator}
\alias{get}
\alias{set}
\alias{get_coordinates}
\alias{get_coordinates-method}
\alias{get_contributions}
\alias{get_contributions-method}
\alias{get_distances}
\alias{get_distances-method}
\alias{get_eigenvalues}
\alias{get_eigenvalues-method}
\alias{get_inertia}
\alias{get_inertia-method}
\alias{dim,CA-method}
\title{Get or Set Parts of an Object}
\usage{
get_coordinates(x, ...)

get_contributions(x, ...)

get_distances(x, ...)

get_eigenvalues(x)

get_inertia(x, ...)

\S4method{dim}{CA}(x)
}
\arguments{
\item{x}{An object from which to get or set element(s) (typically a
\linkS4class{CA} object).}

\item{...}{Currently not used.}
}
\value{
\code{get_*} returns a \code{\link{data.frame}}.

\code{set_*} returns an object of the same sort as \code{x} with the new
values assigned.
}
\description{
Getters and setters to retrieve or set parts of an object.
}
\seealso{
Other mutator: 
\code{\link{subset}()}
}
\author{
N. Frerebeau
}
\concept{mutator}
