% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PCA-class}
\alias{PCA-class}
\alias{.PCA}
\title{PCA Results}
\description{
An S4 class to store the results of a principal components analysis.
}
\section{Slots}{

\describe{
\item{\code{center}}{A \code{\link{numeric}} vector giving the column mean of the
initial dataset (active individuals only).}

\item{\code{standard_deviation}}{A \code{\link{numeric}} vector giving the column
standard deviations of the initial dataset (active individuals only).}
}}

\examples{
## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute principal components analysis
X <- pca(A, scale = TRUE, sup_ind = 8:10, sup_var = 7:10)

## Get row coordinates
get_coordinates(X, margin = 1)

## Get column coordinates
get_coordinates(X, margin = 2)

## Get row contributions
get_contributions(X, margin = 1)

## Get correlations between variables and dimensions
get_correlations(X)

## Get eigenvalues
get_eigenvalues(X)
}
\seealso{
Other multivariate analysis: 
\code{\link{CA-class}},
\code{\link{MultivariateAnalysis-class}},
\code{\link{MultivariateResults-class}},
\code{\link{ca}()},
\code{\link{pca}()},
\code{\link{predict}()}
}
\author{
N. Frerebeau
}
\concept{multivariate analysis}
