% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot.R
\docType{methods}
\name{plot_coordinates}
\alias{plot_coordinates}
\alias{plot_individuals}
\alias{plot_individuals-method}
\alias{plot_variables}
\alias{plot_variables-method}
\alias{plot,CA,missing-method}
\alias{plot,PCA,missing-method}
\alias{plot_individuals,PCA-method}
\alias{plot_variables,PCA-method}
\title{Visualize Factor Map}
\usage{
plot_individuals(object, ...)

plot_variables(object, ...)

\S4method{plot}{CA,missing}(
  x,
  margin = c(1, 2),
  axes = c(1, 2),
  active = TRUE,
  sup = TRUE,
  highlight = NULL,
  group = NULL
)

\S4method{plot}{PCA,missing}(
  x,
  margin = 1,
  axes = c(1, 2),
  active = TRUE,
  sup = TRUE,
  highlight = NULL,
  group = NULL
)

\S4method{plot_individuals}{PCA}(
  object,
  axes = c(1, 2),
  active = TRUE,
  sup = TRUE,
  highlight = NULL,
  group = NULL
)

\S4method{plot_variables}{PCA}(
  object,
  axes = c(1, 2),
  active = TRUE,
  sup = TRUE,
  highlight = NULL,
  group = NULL
)
}
\arguments{
\item{object, x}{A \linkS4class{CA} or \linkS4class{PCA} object.}

\item{...}{Currently not used.}

\item{margin}{A length-one \code{\link{numeric}} vector giving the subscript
which the data will be returned: \code{1} indicates individuals/rows (the
default), \code{2} indicates variables/columns, \code{c(1, 2)} indicates
rows and columns (\code{CA}).}

\item{axes}{A length-two \code{\link{numeric}} vector giving the dimensions
to be plotted.}

\item{active}{A \code{\link{logical}} scalar: should the active
observations be plotted?}

\item{sup}{A \code{\link{logical}} scalar: should the supplementary
observations be plotted?}

\item{highlight}{A \code{\link{character}} string giving XXX.
It must be one of "\code{coordinates}", "\code{contributions}",
"\code{cos2}" or "\code{distances}". Any unambiguous substring can be given.
If \code{NULL} (the default), no highlighting is applied.}

\item{group}{A vector of categories specifying the categorical
variable from which to color the individuals (only used if \code{highlight}
is \code{NULL}; see below).}
}
\description{
Plots factor map.
}
\section{Aesthetic}{

Point shapes and line types are set whether an observation is a
row/individual or a column/variable and is active or supplementary.

Colors are set according to \code{highlight} and \code{group}:
\itemize{
\item{If \code{highlight} is not \code{NULL}, the color gradient will vary
according to the value of the selected parameter.}
\item{If \code{group} is a \code{numeric} vector, the color gradient and
size will vary by the value of \code{group}.}
\item{If \code{group} is not a \code{numeric} vector, the colors will be
mapped to the levels of \code{group}.}
\item{If both are \code{NULL} (the default), then the same rule as for
shapes is used.}
}
}

\examples{
## Load data
library(codex)
data("zuni")

## Compute correspondence analysis
X <- ca(zuni)

## Plot observations
plot(X)
}
\seealso{
Other visualization: 
\code{\link{plot_contributions}()},
\code{\link{plot_eigenvalues}()}
}
\author{
N. Frerebeau
}
\concept{visualization}
