% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot.R
\docType{methods}
\name{plot_eigenvalues}
\alias{plot_eigenvalues}
\alias{plot_eigenvalues-method}
\alias{plot_variance}
\alias{plot_variance-method}
\alias{plot_eigenvalues,MultivariateAnalysis-method}
\alias{plot_variance,MultivariateAnalysis-method}
\title{Visualize Eigenvalues}
\usage{
plot_eigenvalues(object, ...)

plot_variance(object, ...)

\S4method{plot_eigenvalues}{MultivariateAnalysis}(object, fill = "grey30", border = "grey10")

\S4method{plot_variance}{MultivariateAnalysis}(
  object,
  variance = TRUE,
  cumulative = TRUE,
  fill = "grey30",
  border = "grey10",
  colour = "red"
)
}
\arguments{
\item{object}{A \linkS4class{CA} or \linkS4class{PCA} object.}

\item{...}{Currently not used.}

\item{fill, border}{A \code{\link{character}} string specifying the bars
infilling and border colors.}

\item{variance}{A \code{\link{logical}} scalar: should percentages of
variance be plotted instead of eigenvalues?}

\item{cumulative}{A \code{\link{logical}} scalar: should cumulative
percentages of variance be plotted ? Only used if \code{variance} is
\code{TRUE}.}

\item{colour}{A \code{\link{character}} string specifying the line color.}
}
\description{
Plot eigenvalues or variances histogram.
}
\examples{
## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 20)

## Compute correspondence analysis
X <- ca(A, sup_row = 11:15)

## Plot observations
plot_coordinates(X)
}
\seealso{
Other visualization: 
\code{\link{plot_contributions}()},
\code{\link{plot_coordinates}()}
}
\author{
N. Frerebeau
}
\concept{visualization}
