% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ca.R
\docType{methods}
\name{predict}
\alias{predict}
\alias{predict,CA-method}
\title{Predict New Coordinates}
\usage{
\S4method{predict}{CA}(object, data, margin = 1)
}
\arguments{
\item{object}{A \linkS4class{CA} object..}

\item{data}{An object of supplementary points coercible to a
\code{\link{matrix}} for which to compute principal coordinates.}

\item{margin}{A length-one \code{\link{numeric}} vector giving the subscript
which the data will be returned: \code{1} indicates rows (the default),
\code{2} indicates columns.}
}
\value{
A \code{\link{data.frame}} of coordinates.
}
\description{
Operators acting on objects to extract or replace parts.
}
\examples{
## Create count data matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)
B <- matrix(data = sample(12:20, 42, TRUE), nrow = 7, ncol = 6)

## Compute correspondence analysis
X <- ca(A, sup_rows = 8:10, sup_columns = 7:10)

## Predict new row coordinates
predict(X, B, margin = 1)

## Predict new column coordinates
predict(X, B, margin = 2)
}
\seealso{
Other multivariate analysis: 
\code{\link{CA-class}},
\code{\link{correspondence}}
}
\author{
N. Frerebeau
}
\concept{multivariate analysis}
