% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_labels.R
\docType{methods}
\name{viz_labels}
\alias{viz_labels}
\alias{viz_labels-method}
\alias{viz_labels,MultivariateAnalysis-method}
\title{Non-Overlapping Text Labels}
\usage{
viz_labels(x, ...)

\S4method{viz_labels}{MultivariateAnalysis}(
  x,
  margin,
  ...,
  axes = c(1, 2),
  active = TRUE,
  sup = TRUE,
  highlight = NULL,
  top = 10,
  type = "shadow"
)
}
\arguments{
\item{x}{A \code{\linkS4class{CA}}, \code{\linkS4class{MCA}} or \code{\linkS4class{PCA}} object.}

\item{...}{Further \link[graphics:par]{graphical parameters}, particularly,
character expansion, \code{cex} and color, \code{col}.}

\item{margin}{A length-one \code{\link{numeric}} vector giving the subscript
which the data will be returned: \code{1} indicates individuals/rows (the
default), \code{2} indicates variables/columns.}

\item{axes}{A length-two \code{\link{numeric}} vector giving the dimensions to be
plotted.}

\item{active}{A \code{\link{logical}} scalar: should the active observations be
plotted?}

\item{sup}{A \code{\link{logical}} scalar: should the supplementary observations be
plotted?}

\item{highlight}{A vector specifying the information to be highlighted.
If \code{NULL} (the default), no highlighting is applied. If a single \code{character}
string is passed, it must be the name of a categorical variable, or one of
"\code{observation}", "\code{mass}", "\code{sum}", "\code{contribution}" or "\code{cos2}"
(see \code{\link[=augment]{augment()}}).}

\item{top}{An \code{\link{integer}} specifying the number of labels to draw.
Only the labels of the \code{top} \eqn{n} observations will be drawn.
If \code{NULL}, all labels are drawn.}

\item{type}{A \code{\link{character}} string specifying the shape of the field.
It must be one of "\code{text}", "\code{shadow}" or "\code{box}". Any unambiguous substring
can be given.}
}
\description{
Non-Overlapping Text Labels
}
\author{
N. Frerebeau
}
\keyword{internal}
