% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/viz_coordinates.R
\docType{methods}
\name{viz_variables}
\alias{viz_variables}
\alias{viz_variables-method}
\alias{viz_columns}
\alias{viz_columns-method}
\alias{viz_columns,MultivariateAnalysis-method}
\alias{viz_columns,BootstrapCA-method}
\alias{viz_variables,PCA-method}
\alias{viz_variables,BootstrapPCA-method}
\title{Visualize Variables Factor Map}
\usage{
viz_variables(x, ...)

viz_columns(x, ...)

\S4method{viz_columns}{MultivariateAnalysis}(
  x,
  axes = c(1, 2),
  active = TRUE,
  sup = TRUE,
  labels = FALSE,
  map_alpha = NULL,
  map_color = NULL,
  map_shape = NULL,
  map_size = NULL,
  scale_color = NULL,
  scale_shape = NULL,
  scale_size = NULL,
  main = NULL,
  sub = NULL,
  ...
)

\S4method{viz_columns}{BootstrapCA}(x, axes = c(1, 2), scale_color = NULL, scale_shape = NULL, ...)

\S4method{viz_variables}{PCA}(
  x,
  axes = c(1, 2),
  active = TRUE,
  sup = TRUE,
  labels = TRUE,
  map_alpha = NULL,
  map_color = NULL,
  map_linetype = NULL,
  map_size = NULL,
  scale_color = NULL,
  scale_linetype = NULL,
  scale_size = NULL,
  main = NULL,
  sub = NULL,
  ...
)

\S4method{viz_variables}{BootstrapPCA}(x, axes = c(1, 2), scale_color = NULL, scale_shape = NULL, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{CA}} or \code{\linkS4class{PCA}} object.}

\item{...}{Currently not used.}

\item{axes}{A length-two \code{\link{numeric}} vector giving the dimensions to be
plotted.}

\item{active}{A \code{\link{logical}} scalar: should the active observations be
plotted?}

\item{sup}{A \code{\link{logical}} scalar: should the supplementary observations be
plotted?}

\item{labels}{A \code{\link{logical}} scalar: should labels be drawn?}

\item{map_alpha, map_color, map_linetype, map_shape, map_size}{A vector
specifying the information to be highlighted.
It will be mapped to the corresponding aesthetic, according to the
corresponding \code{scale} argument (see examples and vignettes).
If a single \code{character} string is passed, it must be one of "\code{observation}",
"\code{mass}", "\code{sum}", "\code{contribution}" or "\code{cos2}" (see details).
Any unambiguous substring can be given.
If \code{NULL} (the default), no highlighting is applied.}

\item{scale_color}{The colors for points. Multiple colors can be
specified so that each point can be given its own color.}

\item{scale_shape}{A vector of plotting characters or symbols.}

\item{scale_size}{A \code{numeric} vector giving the amount by which plotting
characters and symbols should be scaled relative to the default.}

\item{main}{A \code{\link{character}} string giving a main title for the plot.}

\item{sub}{A \code{\link{character}} string giving a subtitle for the plot.}

\item{scale_linetype}{A vector of line type specification.}
}
\value{
\verb{viz_*()} is called for its side-effects: it results in a graphic
being displayed. Invisibly returns \code{x}.
}
\description{
Plots column/variable principal coordinates.
}
\details{
Available statistics:
\describe{
\item{\code{observation}}{Whether an observation is active or supplementary.}
\item{\code{mass}}{Weight/mass of each observation.}
\item{\code{sum}}{Sum of squared coordinates along \code{axes}.}
\item{\code{contribution}}{Joint contributions to the definition of \code{axes}.}
\item{\code{cos2}}{Joint \eqn{cos^2}{cos2} along \code{axes}.}
}
}
\examples{
## Load data
data("iris")

## Compute principal components analysis
X <- pca(iris, scale = TRUE)

## Plot individuals
viz_individuals(X, map_color = iris$Species, map_shape = iris$Species)

viz_individuals(X, map_color = iris$Sepal.Width, map_size = "cos2")

viz_individuals(X, map_color = "contribution", map_size = "contribution",
                scale_shape = 16)

## Plot variables
viz_variables(X)

viz_variables(X, map_color = "contribution")
}
\seealso{
Other plot methods: 
\code{\link{biplot}()},
\code{\link{screeplot}()},
\code{\link{viz_contributions}()},
\code{\link{viz_individuals}()},
\code{\link{viz_wrap}},
\code{\link{wrap}}
}
\author{
N. Frerebeau
}
\concept{plot methods}
